// src/enums/seller-sort.enum.ts
export enum SellerListingSort {
  NEWEST = 'newest',
  TITLE_ASC = 'title_asc',
  TITLE_DESC = 'title_desc',
  PRICE_LOW_TO_HIGH = 'price_low_to_high',
  PRICE_HIGH_TO_LOW = 'price_high_to_low',
  STOCK_LOW_TO_HIGH = 'stock_low_to_high',
  STOCK_HIGH_TO_LOW = 'stock_high_to_low',
}

export enum SellerOrderSort {
  NEWEST = 'newest',
  OLDEST = 'oldest',
  AMOUNT_HIGH_TO_LOW = 'amount_high_to_low',
  AMOUNT_LOW_TO_HIGH = 'amount_low_to_high',
}